refactor: migrate to BIGINT IDs and enforce clean architecture
- Standardized all entity IDs to Long (BIGINT alignment)
- Refactored controllers to use service layer only
- Introduced service layer across modules
- Implemented WorkOrder state machine with DB-driven transitions
- Added transition logging and validation rules
- Replaced constructors with Lombok builders
- Cleaned imports and removed wildcard usage

+++ backend/hesco/src/main/java/com/lmkr/hesco/user/api/dto/AppUserRequest.java
@@ -9,7 +9,7 @@ public record AppUserRequest(
     @NotBlank String firstName,
     @NotBlank String lastName,
     String contactNumber,
-    @NotNull Short roleId,
+    @NotNull Long roleId,
     Long circleId,
     Long divisionId,
     Long subDivisionId,
